# Node-App/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "version": "1.0.0",
        "title": "APIs Document",
        "description": "Swagger Documentation",
        "termsOfService": "",
        "contact": {
            "name": "Kasun Kulathunga",
            "email": "kasun.kulathunga@gmail.com",
            "url": "https://kasun.kulathunga"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "servers": [{
        "url": "http://localhost:5000/",
        "description": "Local server"
    }],
...
    "paths": {
...
        "/blog/saveComment": {
            "post": {
                "tags": ["Blog"],
                "description": "Save Comment on a Blog Post",
                "operationId": "saveComment",
                "summary": "No Auth Required",
                "security": [],
                "responses": {
                    "200": {
                        "description": "Save Comment on a Blog Post",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "requestBody": {
                    "description": "Please send the _id and Comments array in the body",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Blog"
                            },
                            "examples": {
                                "Add Comments": {
                                    "summary": "Creating a sample blog comment",
                                    "value": {
                                        "_id": "61b6a89cde43ba7ca84dbe8c",
                                        "comments":[
                                            {
                                               "name":"Bob",
                                               "date":"June 25, 2021 05:46:AM",
                                               "comment":"Here's where you'll find this repository's source files. To give your users an idea of what they'll find here, add a description to your repository.",
                                               "reply":[
                                                  {
                                                     "name":"Sharah",
                                                     "date":"June 25, 2021 05:46:AM",
                                                     "comment":"Here's where you'll find this repository's source files. To give your users an idea of what they'll find here, add a description to your repository.",
                                                     "reply":[
                                                        {
                                                           "name":"Jason",
                                                           "date":"June 25, 2021 05:46:AM",
                                                           "comment":"Here's where you'll find this repository's source files. To give your users an idea of what they'll find here, add a description to your repository."
                                                        }
                                                     ]
                                                  }
                                               ]
                                            }]
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "x-codegen-request-body-name": "body"
            }
        }
    },
    "components": {
        "schemas": {
            "Blog": {
                "required": [],
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Sample title"
                    },
                    "content": {
                        "type": "double",
                        "example": "Sample content"
                    },
                    "date": {
                        "type": "integer",
                        "example": 1576281600
                    }
                }
            }
        },
        "securitySchemes": {
            "api_auth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": {
        "api_auth": []
    }
}